skills/instagram-carousel/tools/requirements.txt
# Dependencies for carousel image generation
google-generativeai>=0.8.0
pillow>=10.0.0
